requirements.txt
# LangChain 1.2.0+ (Modern API)
langchain>=1.2.0
langchain-community>=0.4.0
langchain-core>=1.2.0
langchain-groq>=1.1.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
dspy-ai>=2.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
groq>=0.4.0